Public API update part 2
Simplified HoverCar class, moved all API classes to subpackage, added
deprecation warning, removed unused imports, fixed ImportScriptActivity.
API Demo and MCPELauncher-app also were updated.
