--- res/layout/import_patch_confirm.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright (C) 2008 The Android Open Source Project
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-  
-          http://www.apache.org/licenses/LICENSE-2.0
-  
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-
-<!--
-
-  Defines the layout of the splash screen that displays the security
-  settings required for an application and requests the confirmation of the
-  user before it is installed.
--->
-
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_weight="1"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <TextView android:id="@+id/app_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceLarge"
-        android:textColor="?android:attr/textColorPrimary"
-        android:shadowRadius="2"
-        android:singleLine="true"
-        android:gravity="center"
-        android:layout_centerInParent="true"
-        android:paddingRight="6dip"
-        android:paddingLeft="6dip"
-        android:ellipsize="end"/>
-
-    <TextView
-        android:id="@+id/install_confirm_question"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/manage_patches_import_confirm"
-        android:textAppearance="?android:attr/textAppearanceMedium"
-        android:paddingTop="20dip"
-        android:paddingLeft="20dip"
-        android:paddingBottom="20dip"/>
-
-    <!-- OK confirm and cancel buttons.  -->
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        style="@android:style/ButtonBar"
-    >
-        <Button android:id="@+id/ok_button"
-            android:layout_width="0dip"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:filterTouchesWhenObscured="true"
-            android:text="@string/manage_patches_import"
-        />
-
-        <Button android:id="@+id/cancel_button"
-            android:layout_width="0dip"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@android:string/cancel"
-        />
-    </LinearLayout>
-</LinearLayout>
-
